Move logic to de-duplicate extensions into a helper file for shared use
