+++ packages/devtools_app/lib/devtools_app.dart
@@ -4,6 +4,7 @@
 
 export 'src/app.dart';
 export 'src/extensions/extension_service.dart';
+export 'src/extensions/extension_service_helpers.dart';
 export 'src/framework/app_bar.dart';
 export 'src/framework/home_screen.dart';
 export 'src/framework/notifications_view.dart';
