+++ packages/devtools_app/test/extensions/extension_service_test.dart
@@ -171,7 +171,10 @@ void main() {
         StubDevToolsExtensions.barExtension,
         StubDevToolsExtensions.bazExtension,
         StubDevToolsExtensions.someToolExtension,
-      ]..forEach(service.setExtensionIgnored);
+      ];
+      for (final e in extensionsToIgnore) {
+        service.setExtensionIgnored(e, ignore: true);
+      }
       for (final ext in StubDevToolsExtensions.extensions()) {
         expect(
           service.isExtensionIgnored(ext),
@@ -184,98 +187,6 @@ void main() {
       for (final ext in StubDevToolsExtensions.extensions()) {
         expect(service.isExtensionIgnored(ext), false);
       }
-    });
-
-    test('takeLatestExtension returns newer extension', () {
-      expect(
-        takeLatestExtension(
-          StubDevToolsExtensions.barExtension,
-          StubDevToolsExtensions.newerBarExtension,
-        ),
-        StubDevToolsExtensions.newerBarExtension,
-      );
-    });
-
-    test('takeLatestExtension handles parsing errors', () {
-      // Returns 'b' when 'a' has parsing errors.
-      var a = DevToolsExtensionConfig.parse({
-        DevToolsExtensionConfig.nameKey: 'bar',
-        DevToolsExtensionConfig.issueTrackerKey: 'www.google.com',
-        DevToolsExtensionConfig.versionKey: 'this-will-not-parse',
-        DevToolsExtensionConfig.materialIconCodePointKey: 0xe638,
-        DevToolsExtensionConfig.requiresConnectionKey: 'false',
-        DevToolsExtensionConfig.extensionAssetsPathKey: '/absolute/path/to/bar',
-        DevToolsExtensionConfig.devtoolsOptionsUriKey:
-            'file:///path/to/options/file',
-        DevToolsExtensionConfig.isPubliclyHostedKey: 'false',
-        DevToolsExtensionConfig.detectedFromStaticContextKey: 'true',
-      });
-      var b = DevToolsExtensionConfig.parse({
-        DevToolsExtensionConfig.nameKey: 'bar',
-        DevToolsExtensionConfig.issueTrackerKey: 'www.google.com',
-        DevToolsExtensionConfig.versionKey: '2.1.0',
-        DevToolsExtensionConfig.materialIconCodePointKey: 0xe638,
-        DevToolsExtensionConfig.requiresConnectionKey: 'false',
-        DevToolsExtensionConfig.extensionAssetsPathKey: '/absolute/path/to/bar',
-        DevToolsExtensionConfig.devtoolsOptionsUriKey:
-            'file:///path/to/options/file',
-        DevToolsExtensionConfig.isPubliclyHostedKey: 'false',
-        DevToolsExtensionConfig.detectedFromStaticContextKey: 'true',
-      });
-      expect(takeLatestExtension(a, b), b);
-
-      // Returns 'a' when 'b' has parsing errors.
-      a = DevToolsExtensionConfig.parse({
-        DevToolsExtensionConfig.nameKey: 'bar',
-        DevToolsExtensionConfig.issueTrackerKey: 'www.google.com',
-        DevToolsExtensionConfig.versionKey: '2.1.0',
-        DevToolsExtensionConfig.materialIconCodePointKey: 0xe638,
-        DevToolsExtensionConfig.requiresConnectionKey: 'false',
-        DevToolsExtensionConfig.extensionAssetsPathKey: '/absolute/path/to/bar',
-        DevToolsExtensionConfig.devtoolsOptionsUriKey:
-            'file:///path/to/options/file',
-        DevToolsExtensionConfig.isPubliclyHostedKey: 'false',
-        DevToolsExtensionConfig.detectedFromStaticContextKey: 'true',
-      });
-      b = DevToolsExtensionConfig.parse({
-        DevToolsExtensionConfig.nameKey: 'bar',
-        DevToolsExtensionConfig.issueTrackerKey: 'www.google.com',
-        DevToolsExtensionConfig.versionKey: 'this-will-not-parse',
-        DevToolsExtensionConfig.materialIconCodePointKey: 0xe638,
-        DevToolsExtensionConfig.requiresConnectionKey: 'false',
-        DevToolsExtensionConfig.extensionAssetsPathKey: '/path/to/bar',
-        DevToolsExtensionConfig.devtoolsOptionsUriKey: '/path/to/options/file',
-        DevToolsExtensionConfig.isPubliclyHostedKey: 'false',
-        DevToolsExtensionConfig.detectedFromStaticContextKey: 'true',
-      });
-      expect(takeLatestExtension(a, b), a);
-
-      // Returns 'a' when both 'a' and 'b' have parsing errors.
-      a = DevToolsExtensionConfig.parse({
-        DevToolsExtensionConfig.nameKey: 'bar',
-        DevToolsExtensionConfig.issueTrackerKey: 'www.google.com',
-        DevToolsExtensionConfig.versionKey: 'this-will-not-parse',
-        DevToolsExtensionConfig.materialIconCodePointKey: 0xe638,
-        DevToolsExtensionConfig.requiresConnectionKey: 'false',
-        DevToolsExtensionConfig.extensionAssetsPathKey: '/absolute/path/to/bar',
-        DevToolsExtensionConfig.devtoolsOptionsUriKey:
-            'file:///path/to/options/file',
-        DevToolsExtensionConfig.isPubliclyHostedKey: 'false',
-        DevToolsExtensionConfig.detectedFromStaticContextKey: 'true',
-      });
-      b = DevToolsExtensionConfig.parse({
-        DevToolsExtensionConfig.nameKey: 'bar',
-        DevToolsExtensionConfig.issueTrackerKey: 'www.google.com',
-        DevToolsExtensionConfig.versionKey: 'this-will-not-parse',
-        DevToolsExtensionConfig.materialIconCodePointKey: 0xe638,
-        DevToolsExtensionConfig.requiresConnectionKey: 'false',
-        DevToolsExtensionConfig.extensionAssetsPathKey: '/absolute/path/to/bar',
-        DevToolsExtensionConfig.devtoolsOptionsUriKey:
-            'file:///path/to/options/file',
-        DevToolsExtensionConfig.isPubliclyHostedKey: 'false',
-        DevToolsExtensionConfig.detectedFromStaticContextKey: 'true',
-      });
-      expect(takeLatestExtension(a, b), a);
     });
   });
 }
